extraccion de BD/requirements.txt
Flask
numpy
matplotlib
mysql-connector-python
